# carwash-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Wash Admin API Documentation",
        "description": "API documentation for Car Wash Admin backend application (Laravel/Sanctum)",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of all orders",
                "operationId": "9aef6b25b778c1a661d160b65438b7ec",
                "parameters": [
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "description": "Filter by status (e.g., new, in_progress)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order with items (in a transaction)",
                "operationId": "047e9e9d51695501706383f331999ad1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "description": "Optional vehicle ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "manager_id": {
                                        "description": "Optional manager ID",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Mijoz 1 soatda keladi"
                                    },
                                    "items": {
                                        "description": "List of services for the order (Order Items)",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "service_id",
                                                "qty"
                                            ],
                                            "properties": {
                                                "service_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "worker_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get details for a specific order",
                "operationId": "7a42445ea1d3933c6ce0025b0239ae1f",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an existing order (Hozircha faqat status va asosiy ma'lumotlar)",
                "operationId": "c66ff7d9c25c9d910e0f70751f9ce8e8",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "in_progress",
                                            "done",
                                            "paid",
                                            "canceled"
                                        ],
                                        "example": "done"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Buyurtma yakunlandi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete a specific order",
                "operationId": "7f17a66646c70db839f68fc4a7b06c60",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserResource": {
                "title": "User Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Alijon Valiyev"
                    },
                    "phone": {
                        "type": "string",
                        "example": "998901234567"
                    },
                    "email": {
                        "type": "string",
                        "example": "ali@example.com"
                    },
                    "tenantId": {
                        "type": "integer",
                        "example": 1
                    },
                    "role": {
                        "type": "string",
                        "example": "manager"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "OrderResource": {
                "title": "Order Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ticketNo": {
                        "type": "string",
                        "example": "T-1701234567"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "in_progress",
                            "done",
                            "paid",
                            "canceled"
                        ],
                        "example": "new"
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": "50000.00"
                    },
                    "manager": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter Sanctum token for authentication (Bearer Token)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}